QuantumPulse/requirements.txt
# Core API
fastapi
uvicorn[standard]
pydantic
openai

# Messaging
pulsar-client

# Configuration
pyyaml

# In-Memory Data Grid / Cache
pyignite

# Stream Processing
apache-flink

# MLOps / Fine-Tuning
transformers
datasets
accelerate
huggingface_hub
torch
pennylane

# Utilities
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
prometheus-client
structlog
